src/main/webapp/app/entities/relationship/route/relationship-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IRelationship } from '../relationship.model';
import { RelationshipService } from '../service/relationship.service';

import { RelationshipRoutingResolveService } from './relationship-routing-resolve.service';

describe('Relationship routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: RelationshipRoutingResolveService;
  let service: RelationshipService;
  let resultRelationship: IRelationship | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(RelationshipRoutingResolveService);
    service = TestBed.inject(RelationshipService);
    resultRelationship = undefined;
  });

  describe('resolve', () => {
    it('should return IRelationship returned by find', () => {
      // GIVEN
      service.find = jest.fn(relationshipId => of(new HttpResponse({ body: { relationshipId } })));
      mockActivatedRouteSnapshot.params = { relationshipId: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRelationship = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultRelationship).toEqual({ relationshipId: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRelationship = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultRelationship).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IRelationship>({ body: null })));
      mockActivatedRouteSnapshot.params = { relationshipId: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRelationship = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultRelationship).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
